requirements.txt
pandas
numpy
scikit-learn
xgboost
tensorflow
fastapi
uvicorn[standard]
streamlit
langchain
langchain-google-genai
google-generativeai
matplotlib
seaborn
joblib
requests
plotly
python-dotenv
pydantic
